Yet another way to fix issue #1 (EJS, swig generators)
'jsxgettext' variable will contain this:
{
  parse: [Function: parse],
  generate: [Function: gen],
  generateFromEJS: [Function: EJS],
  generateFromJinja: [Function: Jinja],
  generateFromHandlebars: [Function: Handlebars],
  generateFromJade: [Function: Jade],
  generateFromswig: [Function: swig]
}

As you can see the 'generateFrom' pattern is not consistent, the 'ejs' is fully upper case and 'swig' is fully lower case.
A switch-case at 'generator' variable population is implemented to workaround this inconsistency.
This workaround doesn't duplicate the generator function like maxnachlinger's workaround.
This workaround also provides a case insensitive matching of parser names in 'parsers' option of extract function.